Fix dijkstra + a* fastest

+++ be-graphes-algos/src/main/java/org/insa/graphs/algorithm/shortestpath/LabelStar.java
@@ -4,25 +4,24 @@ import org.insa.graphs.model.Arc;
 import org.insa.graphs.model.Node;
 
 public class LabelStar extends Label {
-
 	
-	private float estimatedCost;
+	private double estimatedCost;
 	
-	public LabelStar(Node sc, boolean m, float c, Arc p, float ec) {
+	public LabelStar(Node sc, boolean m, double c, Arc p, double ec) {
 		super(sc, m, c, p);
 		this.estimatedCost = ec;
 	}
 	
-	public float getEstimatedCost() {
+	public double getEstimatedCost() {
 		return this.estimatedCost;
 	}
 	
-	public void setEstimatedCost(float value) {
+	public void setEstimatedCost(double value) {
 		this.estimatedCost = value;
 	}
 	
 	@Override
-	public float getTotalCost() {
+	public double getTotalCost() {
 		return this.estimatedCost + this.cost;
 	}
 }
